Refactor using showWarningMessage
