+++ src/commands/createFunctionApp/stacks/FunctionAppStackStep.ts
@@ -4,9 +4,11 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { setLocationsTask, SiteOSStep, WebsiteOS } from '@microsoft/vscode-azext-azureappservice';
-import { AzureWizardPromptStep, IAzureQuickPickItem, IWizardOptions } from '@microsoft/vscode-azext-utils';
+import { AzureWizardPromptStep, DialogResponses, IAzureQuickPickItem, IWizardOptions, openUrl } from '@microsoft/vscode-azext-utils';
+import { MessageItem } from 'vscode';
 import { getMajorVersion, promptForFuncVersion } from '../../../FuncVersion';
 import { localize } from '../../../localize';
+import { getWorkspaceSetting, updateGlobalSetting } from '../../../vsCodeConfig/settings';
 import { FullFunctionAppStack, IFunctionAppWizardContext } from '../IFunctionAppWizardContext';
 import { getStackPicks } from './getStackPicks';
 
@@ -20,6 +22,29 @@ export class FunctionAppStackStep extends AzureWizardPromptStep<IFunctionAppWiza
             if (!result) {
                 context.version = await promptForFuncVersion(context);
             } else {
+                const endofLifeDate = result.minorVersion.stackSettings.linuxRuntimeSettings?.endOfLifeDate;
+                const sixMonthsFromNow = new Date();
+                sixMonthsFromNow.setMonth(sixMonthsFromNow.getMonth() + 6);
+                const compareDates = (eolDate: string, sixMonthsDate: Date) => {
+                    const endOfLife = new Date(eolDate).getTime();
+                    const sixMonths = new Date(sixMonthsDate).getTime();
+                    return endOfLife >= sixMonths;
+                }
+                const settingKey: string = 'endOfLifeWarning';
+                if (getWorkspaceSetting<boolean>('endOfLifeWarning')) {
+                    if (endofLifeDate) {
+                        if (!compareDates(endofLifeDate, sixMonthsFromNow)) {
+                            const message = localize('endOfLife', "The chosen runtime stack has an end of support deadline coming up. After the deadline, function apps can be created and deployed, and existing apps continue to run. However, your apps won't be eligible for new features, security patches, performance optimizations, and support until you upgrade them");
+                            const result: MessageItem = await context.ui.showWarningMessage(message, DialogResponses.learnMore, DialogResponses.dontWarnAgain);
+                            if (result === DialogResponses.learnMore) {
+                                await openUrl('https://learn.microsoft.com/en-us/azure/azure-functions/functions-versions?tabs=azure-cli%2Cwindows%2Cin-process%2Cv4&pivots=programming-language-csharp');
+                            }
+                            else if (result === DialogResponses.dontWarnAgain) {
+                                await updateGlobalSetting(settingKey, false);
+                            }
+                        }
+                    }
+                }
                 break;
             }
         }
